Move some unit tests to the experiment module
